Updated default_rate_analysis.sql

+++ src/database/queries/default_rate_analysis.sql
@@ -31,14 +31,14 @@ SELECT
     grade,
     COUNT(*)                                            AS total_mature_loans,
     SUM(is_default)                                     AS defaulted_loans,
-    ROUND(100.0 * SUM(is_default) / COUNT(*), 2)       AS default_rate_pct,
-    ROUND(AVG(loan_amnt), 2)                            AS avg_loan_amount,
-    ROUND(AVG(int_rate), 2)                             AS avg_interest_rate,
-    ROUND(AVG((fico_range_low + fico_range_high) / 2.0), 1) AS avg_fico_score,
-    ROUND(AVG(annual_inc), 2)                           AS avg_annual_income,
-    ROUND(AVG(dti), 2)                                  AS avg_dti,
-    ROUND(AVG(emp_length), 1)                           AS avg_emp_length_years,
-    ROUND(AVG(credit_history_months), 1)                AS avg_credit_history_months
+    ROUND(100.0 * SUM(is_default) / COUNT(*)::numeric, 2) AS default_rate_pct,
+    ROUND(AVG(loan_amnt)::numeric, 2)                   AS avg_loan_amount,
+    ROUND(AVG(int_rate)::numeric, 2)                    AS avg_interest_rate,
+    ROUND(AVG((fico_range_low + fico_range_high) / 2.0)::numeric, 1) AS avg_fico_score,
+    ROUND(AVG(annual_inc)::numeric, 2)                  AS avg_annual_income,
+    ROUND(AVG(dti)::numeric, 2)                         AS avg_dti,
+    ROUND(AVG(emp_length)::numeric, 1)                  AS avg_emp_length_years,
+    ROUND(AVG(credit_history_months)::numeric, 1)       AS avg_credit_history_months
 FROM labeled
 GROUP BY grade
 ORDER BY grade;
